lint

+++ tests/blockchain/blockchain_test_utils.py
@@ -7,7 +7,7 @@ from chia.consensus.blockchain import AddBlockResult, Blockchain
 from chia.consensus.multiprocess_validation import PreValidationResult
 from chia.types.full_block import FullBlock
 from chia.util.errors import Err
-from chia.util.ints import uint64
+from chia.util.ints import uint64, uint32
 
 
 async def check_block_store_invariant(bc: Blockchain):
@@ -60,7 +60,7 @@ async def _validate_and_add_block(
     else:
         fork_height = None
         if fork_info is not None:
-            fork_height = fork_info.fork_height
+            fork_height = uint32(fork_info.fork_height)
         # Do not change this, validate_signatures must be False
         pre_validation_results: List[PreValidationResult] = await blockchain.pre_validate_blocks_multiprocessing(
             [block], {}, fork_height=fork_height, validate_signatures=False
